Implement RN bridge
