+++ src/PrimerNative.ts
@@ -1,0 +1,2 @@
+import { NativeModules } from 'react-native';
+export const { PrimerNative: NativeModule } = NativeModules;
